# [Crypto] performance improvement of SSWU map to curve