--- a/crypto/bls_test.go
+++ b/crypto/bls_test.go
@@ -1,3 +1,4 @@
+//go:build relic
 // +build relic
 
 package crypto