The crypto library has been using Relic's map-to-curve implementation of [SSWU algorithm](https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-hash-to-curve-11#section-6.6.2).

Specific optimizations could be implemented in the case of mapping to BLS12 381 (A == 0, and p mod 4 = 3) to improve performance. The optimizations are detailed in the IETF draft ([optimized SSWU](https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-hash-to-curve-14#appendix-F.2) and [square root ratio](https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-hash-to-curve-14#appendix-F.2.1.2)).

This PR implements the optimized SSWU for BLS12 381 by updating an older implementation of SSWU (map [coefficient z](https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-hash-to-curve-14#appendix-H.2) updated from -1 to 11 as per the IETF draft) as well as optimizing the computation steps. 

Other updates:
- remove hardcoded constants and rely on Relic for SSWU and BLS12 381 constants (z, isogeny map, prime field).
- cache extra constants required for SSWU (isogeny map constants, Montgomery constants)
- separate bench function for scalar multiplication of generator and generic points
- add a new test vector from the IETF draft 

Bench results on my laptop show about 35% improvement:
```
name       old time/op  new time/op  delta
MapToG1-8   364µs ± 1%   232µs ± 1%  -36.14%  (p=0.000 n=30+29)
```